requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bottle==0.12.13
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
decorator==4.2.1
Django==2.0.1
django-extensions==1.9.9
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
lxml==4.1.1
mecab-python3==0.7
parsel==1.3.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.5.0
pytz==2017.3
queuelib==1.4.2
requests==2.18.4
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.4
urllib3==1.22
uWSGI==2.0.16
w3lib==1.18.0
wcwidth==0.1.7
zope.interface==4.4.3
